Changed file location and edited asset locations also
